Prune unneeded data as driver runs

+++ R/utils.R
@@ -218,31 +218,35 @@ find_csv_file <- function(filename, optional, quiet = FALSE) {
 #' a lot bigger, owing to the large number of digits we have to write.
 #'
 #' @param chunkdata Named list of tibbles (data frames) to write
-#' @param write_inputs Write data that were read as inputs, not computed?
+#' @param write_inputs Write data that were read as inputs, not computed? Logical
+#' @param create_dirs Create directory if necessary, and delete contents? Logical
 #' @param outputs_dir Directory to save data into
 #' @param xml_dir Directory to save XML results into
 #' @importFrom assertthat assert_that
-save_chunkdata <- function(chunkdata, write_inputs = FALSE, outputs_dir =
-                             OUTPUTS_DIR, xml_dir = XML_DIR) {
+save_chunkdata <- function(chunkdata, write_inputs = FALSE, create_dirs = FALSE,
+                           outputs_dir = OUTPUTS_DIR, xml_dir = XML_DIR) {
   assert_that(is_data_list(chunkdata))
-  assert_that(!is.null(names(chunkdata)))
   assert_that(is.logical(write_inputs))
+  assert_that(is.logical(create_dirs))
   assert_that(is.character(outputs_dir))
 
   # Create directory if necessary, and remove any previous outputs
-  dir.create(outputs_dir, showWarnings = FALSE, recursive = TRUE)
-  unlink(file.path(outputs_dir, "*.csv"))
-  dir.create(xml_dir, showWarnings = FALSE, recursive = TRUE)
-  unlink(file.path(xml_dir, "*.xml"))
+  if(create_dirs) {
+    dir.create(outputs_dir, showWarnings = FALSE, recursive = TRUE)
+    unlink(file.path(outputs_dir, "*.csv"))
+    dir.create(xml_dir, showWarnings = FALSE, recursive = TRUE)
+    unlink(file.path(xml_dir, "*.xml"))
+  }
 
   for(cn in names(chunkdata)) {
-    cd <- chunkdata[[cn]]
+    cd <- get_data(chunkdata, cn)
+    if(is.null(cd)) next   # optional file that wasn't found
+
     if(FLAG_XML %in% get_flags(cd)) {
       # TODO: worry about absolute paths?
       cd$xml_file <- file.path(xml_dir, cd$xml_file)
       run_xml_conversion(cd)
-    } else if(!isTRUE(identical(NA, cd))) {   # NA means an optional file that wasn't found
-
+    } else {
       fqfn <- file.path(outputs_dir, paste0(cn, ".csv"))
       suppressWarnings(file.remove(fqfn))
 
@@ -370,6 +374,18 @@ chunk_inputs <- function(chunks = find_chunks()$name) {
 }
 
 
+#' inputs_of
+#'
+#' Convenience function for getting the inputs of one or more chunks
+#'
+#' @param chunks Names of chunks, character
+#' @return Character vector of inputs.
+#' @export
+inputs_of <- function(chunks) {
+  if(is.null(chunks) || chunks == "") return(NULL)
+  chunk_inputs(chunks)$input
+}
+
 #' chunk_outputs
 #'
 #' List all chunk outputs.
@@ -399,6 +415,17 @@ chunk_outputs <- function(chunks = find_chunks()$name) {
   dplyr::bind_rows(chunkoutputs)
 }
 
+#' outputs_of
+#'
+#' Convenience function for getting the outputs of one or more chunks
+#'
+#' @param chunks Names of chunks, character
+#' @return Character vector of inputs.
+#' @export
+outputs_of <- function(chunks) {
+  if(is.null(chunks) || chunks == "") return(NULL)
+  chunk_outputs(chunks)$output
+}
 
 #' screen_forbidden
 #'
